requirements.txt
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
google-api-python-client==1.7.9
google-auth==2.23.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.22.0
idna==3.4
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
uritemplate==3.0.1
urllib3==1.25.11
